Rename app "core" into "gsl_core"
...I hope it's the only time I must trash the whole database
