--- core/.github/PULL_REQUEST_TEMPLATE.md
@@ -1,21 +1,0 @@
-## 🎯 Objectif
-
-_Un résumé de l’objectif de la PR_
-
-_Précisez si c’est un bugfix, une nouvelle fonctionnalité… Expliquez rapidement le gain « fonctionnel » de vos modifications._
-
-## 🔍 Implémentation
-
-- _Une liste des modifications_
-
-## ⚠️ Informations supplémentaires
-
-_(optionnel) Documentation, commandes à lancer, variables d’environnement, etc._
-
-## 🏕 Amélioration continue
-
-- _(optionnel) Une liste d’autres modifications pas en lien direct avec la PR_
-
-## 🖼️ Images
-
-_(optionnel) Une ou plusieurs captures d’écran, si pertinent_
