+++ gsl_core/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from core.models import Collegue
+from gsl_core.models import Collegue
 
 admin.site.register(Collegue, UserAdmin)
