+++ gsl_core/templates/blocks/footer.html
